refactor(di): move inversify types to infrastructure/di folder
- Relocate TYPES constant from src/types/types.ts to src/infrastructure/di/types.ts
- Update all import paths across the application to reference the new location
- Improve code organization by placing DI related configuration in dedicated infrastructure folder
