+++ src/applications/usecases/record/create-substitution.usecase.ts
@@ -1,5 +1,5 @@
 import { injectable, inject } from "inversify";
-import { TYPES } from "@/types/types";
+import { TYPES } from "@/infrastructure/di/types";
 import type { IRecordRepository } from "@/applications/repositories/record.repository.interface";
 import type { IAuthenticationService } from "@/applications/services/auth/authentication.service.interface";
 import type { IAuthorizationService } from "@/applications/services/auth/authorization.service.interface";
