+++ src/infrastructure/services/auth/authentication.service.ts
@@ -1,5 +1,5 @@
 import { injectable, inject } from "inversify";
-import { TYPES } from "@/types/types";
+import { TYPES } from "@/infrastructure/di/types";
 import { IAuthenticationService } from "@/applications/services/auth/authentication.service.interface";
 import type { IUserRepository } from "@/applications/repositories/user.repository.interface";
 import { User } from "@/entities/user";
@@ -10,7 +10,7 @@ export class AuthenticationService implements IAuthenticationService {
   constructor(
     @inject(TYPES.UserRepository) private userRepository: IUserRepository
   ) {}
-  
+
   async verifySession(): Promise<User | undefined> {
     const session = await auth();
     if (!session) throw new Error("Invalid session");
